# docs/swagger.yaml
basePath: /
definitions:
...
  domain.CreateReviewRequest:
    properties:
      comment:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      target_id:
        type: string
    required:
    - comment
    - rating
    - target_id
    type: object
...
  domain.Review:
    properties:
      author_id:
        type: string
      author_name:
        type: string
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      rating:
        type: number
      target_id:
        type: string
    type: object
  domain.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
      phone:
        minLength: 6
        type: string
      surname:
        minLength: 3
        type: string
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_admin:
        type: boolean
      name:
        type: string
      password:
        description: не возвращаем хеш пароля наружу
        type: string
      phone:
        type: string
      rating:
        type: number
      surname:
        type: string
    type: object
...
info:
  contact: {}
  description: API для виртуальной доски объявлений между студентами
  title: StudShare API
  version: "1.0"
paths:
...
  /reviews:
    delete:
      parameters:
      - description: Review ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: missing review ID
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: review not found
          schema:
            type: string
        "500":
          description: failed to delete review
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить свой отзыв
      tags:
      - reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: Отзыв
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: invalid request body or validation failed
          schema:
            type: string
        "403":
          description: cannot leave a review for yourself
          schema:
            type: string
        "409":
          description: review already exists
          schema:
            type: string
        "500":
          description: failed to add review
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Оставить отзыв
      tags:
      - reviews
  /reviews/user:
    get:
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Review'
            type: array
        "400":
          description: missing user ID
          schema:
            type: string
        "500":
          description: failed to get reviews
          schema:
            type: string
      summary: Получить отзывы о пользователе
      tags:
      - reviews
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить свой профиль
      tags:
      - users
  /users/profile:
    get:
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "404":
          description: User not found
          schema:
            type: string
      summary: Получить профиль пользователя по ID
      tags:
      - users
  /users/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Данные для обновления
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid request or no changes
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Error updating user
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить свой профиль
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"